Moved convert() method from ViewController to Presenter
